+++ app/controllers/workbaskets/base_controller.rb
@@ -44,7 +44,7 @@ module Workbaskets
     end
 
     expose(:form) do
-      "Workbaskets::#{sub_klass}::Form".constantize.new(
+      "Workbaskets::#{sub_klass}Form".constantize.new(
         Measure.new
       )
     end
