+++ app/forms/workbaskets/create_measures_form.rb
@@ -1,0 +1,6 @@
+module Workbaskets
+  class CreateMeasuresForm < ::Workbaskets::BaseForm
+
+    attr_accessor :workbasket_name
+  end
+end
