+++ app/views/workbaskets/create_measures/steps/_duties_conditions_footnotes.html.slim
@@ -1,6 +1,6 @@
 = render "workbaskets/create_measures/steps/duties_conditions_footnotes/duty_expression", f: f
-= render "workbaskets/create_measures/steps/duties_conditions_footnotes/standart_import_value", f: f
-= render "workbaskets/create_measures/steps/duties_conditions_footnotes/reference_price", f: f
-= render "workbaskets/create_measures/steps/duties_conditions_footnotes/unit_price", f: f
-= render "workbaskets/create_measures/steps/duties_conditions_footnotes/conditions", f: f
-= render "workbaskets/create_measures/steps/duties_conditions_footnotes/footnotes", f: f
+= render "workbaskets/shared/steps/duties_conditions_footnotes/standart_import_value", f: f
+= render "workbaskets/shared/steps/duties_conditions_footnotes/reference_price", f: f
+= render "workbaskets/shared/steps/duties_conditions_footnotes/unit_price", f: f
+= render "workbaskets/shared/steps/duties_conditions_footnotes/conditions", f: f
+= render "workbaskets/shared/steps/duties_conditions_footnotes/footnotes", f: f
