+++ app/views/workbaskets/create_measures/steps/_main.html.slim
@@ -1,10 +1,10 @@
 = render "workbaskets/create_measures/steps/main/start_end_dates", f: f
-= render "workbaskets/create_measures/steps/main/regulation", f: f
-= render "workbaskets/create_measures/steps/main/measure_type", f: f
+= render "workbaskets/shared/steps/main/regulation", f: f
+= render "workbaskets/shared/steps/main/measure_type", f: f
 = render "workbaskets/create_measures/steps/main/workbasket_name", f: f
-= render "workbaskets/create_measures/steps/main/operation_date", f: f
-= render "workbaskets/create_measures/steps/main/goods_commodity_codes", f: f
-= render "workbaskets/create_measures/steps/main/additional_codes", f: f
-= render "workbaskets/create_measures/steps/main/reduction_indicator", f: f
-= render "workbaskets/create_measures/steps/main/geographical_area", f: f
+= render "workbaskets/shared/steps/main/operation_date", f: f
+= render "workbaskets/shared/steps/main/goods_commodity_codes", f: f
+= render "workbaskets/shared/steps/main/additional_codes", f: f
+= render "workbaskets/shared/steps/main/reduction_indicator", f: f
+= render "workbaskets/shared/steps/main/geographical_area", f: f
