+++ app/views/workbaskets/create_measures/steps/duties_conditions_footnotes/_duty_expression.html.slim
@@ -1,4 +1,3 @@
-/ fieldset v-if="showDuties"
 fieldset
   h3.heading-medium
     | What duties will apply these measures?
