+++ app/views/workbaskets/create_quota/_custom_errors_block.html.slim
@@ -1,0 +1,23 @@
+.js-workbasket-errors-container.js-custom-errors-block.alert.alert-danger.hidden
+  h3.measure-form-errors-header
+    | Progress successfuly saved!
+    br
+    | Please, review the following errors!
+
+  .js-workbasket-custom-errors.hidden data-errors-container="measure"
+    h3.create-measures-custom-errors-header
+      | General:
+
+    ul
+
+  .js-workbasket-custom-errors.hidden data-errors-container="conditions"
+    h3.create-measures-custom-errors-header
+      | Conditions:
+
+    ul
+
+  .js-workbasket-custom-errors.hidden data-errors-container="footnotes"
+    h3.create-measures-custom-errors-header
+      | Footnotes:
+
+    ul
