+++ app/views/workbaskets/create_quota/_form.html.slim
@@ -1,7 +1,7 @@
 - cache [ "measure_form_js_variables", expires_in: 8.hours ] do
   = render "js_variables"
 
-= simple_form_for form, url: create_measure_url(id: workbasket.id),
+= simple_form_for form, url: create_quotum_url(id: workbasket.id),
                         html: { \
                           class: "measure-form create-measures-v2",
                           method: :put, data: { \
@@ -10,11 +10,11 @@
                           } \
                         } do |f|
 
-  = render "workbaskets/create_measures/steps/#{params[:step].to_s}", f: f
+  = render "workbaskets/create_quota/steps/#{params[:step].to_s}", f: f
   = render "workbaskets/shared/success_message", f: f
   = render "workbaskets/shared/errors_container", f: f
 
-  - if current_step == "duties_conditions_footnotes"
-    = render "workbaskets/create_measures/custom_errors_block", f: f
+  - if current_step == "conditions_footnotes"
+    = render "workbaskets/create_quota/custom_errors_block", f: f
 
   = render "workbaskets/shared/action_links", f: f
