+++ app/views/workbaskets/shared/steps/duties_conditions_footnotes/_conditions.html.slim
