+++ app/views/workbaskets/shared/steps/duties_conditions_footnotes/_footnotes.html.slim
