+++ app/views/workbaskets/shared/steps/main/_operation_date.html.slim
