+++ app/views/workbaskets/shared/steps/main/_reduction_indicator.html.slim
