+++ app/views/workbaskets/shared/steps/main/_regulation.html.slim
