+++ app/views/workbaskets/shared/steps/review_and_submit/_table.html.slim
